debian/gridvis-server.postinst: adding postinst script
The package 'gridvis-server' isn't hardly needing the package
'gridvis-desktop' to work. To let the web services basically work the
package need the directory /var/lib/GridVisProjects and write access to
it.

The suggested way is to have the package 'gridvis-desktop' installed
that will place the project files in /var/lib/GridVisProjects that are
readed by the 'gridvis-server' package. To let the 'gridvis-server'
package work also without the 'gridvis-desktop' package we have to
ensure that we create this directory with the needed rights while the
installation if the 'gridvis-desktop' package isn't installed before.